( 7/16) Unify emissions testing & memo lib in modo agno
Because of how frequently tests test emissions and use memoization, it
made sense to unify (finally) such things into a generic “common” module
that tests can import for these common resources.

At first we went to unify the many different similar but different ways
we do `listener, emissions = …` in tests. It was definitely a pattern,
and as it stood there was tons of redundancy in re-implementing the same
thing over and over.

This grew in to a desire to do a blind rewrite of
“listener via expectations”, because (as it proved to be true) we
thought we could clean it up a lot and make its implementation shorter
and clearer. (Down from a 282 line file, it’s now about 75 lines, and
now alongside the other two major emission testing functions).

Some small memorization functions have independent, semi-redundant
implementations across the “test support” and “production” sides of [ma].
This is intentional, to avoid a thin, trivial dependency on small funcs.