2/4: Rewrite “expect STDs” & absorb a related library
As is our custom, we went to change or fix something here and became
compelled to clean it up, simplify it, and add features to it. It now
weights not as heavily on the conscience to use it, and so we found it
was able to replace some other fellows.