feat(kiss-rdb:md): support custom identifier class injection ((1240))
Discussion: We knew that adding support for custom identifier classes
would break things in data pipes sync

They key “paradigm shift” that occurred as a downstream result of this
change is that a “near sync keyer builder” is only a concern of syncing
(its flat map implementation), not of the rewriting collection

So now, if a syncing client wants to sync its far collection into a
kiss-rdb near collection, it injects its flat map into the sync agent
and it knows the sync agent will feed each *entity* into the flat map.
The syncing client is the one that must deal with doing any near key
generation. But it can do all of this (the order determination etc)
without bothering the near collection rewriter

The really good news is that what was once two separate things (editing
collections and syncing collections) is now one unified thing (and has
been for a few weeks), and also it’s modular enough to accommodate a
custom identifier class or not; all of this uses the same relatively
small codebase built around flat maps

Also:
- Started using this https://www.conventionalcommits.org/en/v1.0.0/

(times)
  09-22 23:43  begin inject custom identifier class into kss
        01:56  we were moving really slow and the test isn't written but
               we think we have identifier injection done
        02:29  take a rest
  09-23 09:30  resume. trying to figure out how soon to test edits with
               injected idens. we decide to go big:
               inj ID in one file, after RDCU
        11:22  passed first good, very thorough test case (no CDU yet)
        12:55  nap
        14:30  resume
        17:36  finished second test case about collection out of order
               (using custom order). break for zumba
  09-24 09:36  begin the hard part: un-breaking syncing
        09:51  begin want to away near keyer
               The hypothesis is that we can get rid of near keyer and
               defer all such work to the injected flat map, which just
               receives entities
        14:15  all green in universe
.