refactor: unify “dangerous memoize in child classes” ((1249))
Discussion: when we made the “old way” we did it with an overabundance
of caution. For some reason we thought all these extra parameters were
necessary. Perhaps we hadn’t thought of using `func.__name__`. At the
time we introduced the newer version, we thought it might be functionally
different; that is, that it was suitable for a different category of
use cases. However, after time we realized we could just keep it simple
and always use the second form. (We suspected this at the time of
creation because of that ugly name ending in `_2`.)

time: 12:11-12:38