refactor: synthesize new “text-lib” module (2/2) ((1299))
Our near-term objective is to implement the “select” operation for
collections. We observed that “kiss-rdb” felt overcrowded with both
collection implementations *and* higher-operations that act on those
collections. Since this was the mandate of “data-pipes” was to house
these higher-level facilities, we decided to move several of the
CLI-level operations from kiss to dp: select, filter-by-tags,
convert-collection.

We were DELIGHTED that we could refactor some of these CLI functions
off of click, which cleaned up the test code a lot.

Unify several ad-hoc constructions of CLI end state and abstract it up
into “expect STDs”

In the process of trying to re-house and streamline the CLI-level
unit tests, we decided it was time to go ahead and break out “text-lib”,
a support module for producing and processing text.

Re-house unindent and rotating buffer (so the oxfords) into under the
same roof as the oldest related participant here: word wrap.

(times)
  10-06 10:05  we decide that for now, what we are doing is
               moving all interesting command (2 of them) to data-pipes:
               "select" and "filter-by-tags”. or maybe text-lib now
        10:23  amazingly, already 2 test files passing in the new home
        12:33  moved liked four modules
  10-07 16:47  next day, done