refactor: begin “magnetic field” collection adapters (1/2) (952)
Freeform discussion:

The main objective it to allow all single-file-based (line-based)
format adapters to get the big list of capabilities “for free” if they
all define themselves in the prescribed way.

A secondary objective is that we get away from thinking that we need
to build a whole, full-featured collection façade for every invocation
of every operation on every collection. Some collections have no edit
capabilities at all (producer scripts); many collections don’t need
edit facilities for the particular operation of the invocation.

A third but equally important objective is to embrace proper
resource management everywhere, so we close files as soon as we are
done with them.

Our experimental solution to this is that rather than collection
adapters defining a single “collection implementation” that they return,
they return an object that produces (or doesn’t produce) functions
*on request* (a “functionser”).

We absolutely don’t want storage adapters to have to back-depend on
kiss rdb. We want their top-level specification to be purely
“declarative” and not “imperative”.

You can pass the number of tests reflected in the
above number by deleting the filetrees named in ‘TEMPORARY_NOT_THESE’

(times will be in the next commit)