feature(data-pipes): select (minimally illustrative) ((1366))
Discussion/Contextualization:

`select` (as we have implemented it here) is so minimal, functionally,
that it’s almost embarrassing it exists as its own thing. BUT:
it frontiers some core principles and properties:

- the data pipes “set theory”: chain functions that compose collections
  from other collections.. Be able to do this equally well across
  processes (through stdout) and within a process

The reason we have done this minimally simple composable function first
is so we can focus on these target properties more than the intricacies
of implementing the function.

Also:
- Universally, `core_attributes` not `core_attributes_dictionary`.
  The fact that it’s a dictionary is an implementation detail.
.