feat(data-pipes): introduce pipelines ((1375))
Three Systems

Over the years, we had come to rely (quite heavily at times)
on our graph viz graphs to aid in development: to aid in (for example)
which things to work on next by making dependency graphs, and using the
nodes in the graph as a “punchlist” to mark of what was completed and
what remained.

Separately we utilize our markdown lists to track open issues and other
..things we want to track.

Separately (still) we rely heavily on our “.stack” file to be (yet
again) a sort of “punchlist” for a variety of kinds of items, from
high-level milestones to very short term scratch-space checklist tasks.


Compare & Contrast the Systems

Each of these systems has its various strengths and weaknesses: the
“.stack” file system is simple and reliable: its implementation code
never breaks, it’s quick and easy to add items to it; it only has two
commands (for practical purposes) with immediately transparent
interfaces. Its main disadvantage is that it doesn’t have all the
features markdown tables have.

Markdown tables have the benefit of being integrated in with other
documentation, looking “pretty” (for definitions of) when generated,
but still being a simple, readable text-based format. Furthermore this
format avails itself well to a tagging system where nodes in the table
can reference each other. The main disadvantage is that the application
code is not yet robust & reliable. Also you have to switch tables
(files) because there are multiple of them (which is also a feature).

Finally, graph viz files are great at what they do, but hand-writing
them can get extremely cumbersome and brittle, and such an approach is
just screaming to be integrated with one or more of the systems above.


The Evergreen Dream

So putting it all together: an evergreen dream we’ve had going back to
forever has been to exploit the “isomorphisms” between these systems
to perhaps unify some of these behavior sets across them.

One (perhaps over-idealized) imagined scenario is that we only use
markdown tables, and it has the robustness of the “.stack” system and
the utility of generating “.dot” files for you. (And in fact we have
done similar experiments with things like letting our “.stack” files
express dependencies, ending lines with expressions like “(A)” or
“(A->B)”).


The stage is set. the plot thickens:

About a month ago (on October 5), we decided we wanted to “absorb” one
of our graph-viz files into a markdown table, so that the official home
for the data would be in the table, and it would generate the graph
mostly faithfully.

Here’s where it gets extra goofy: We have many possible candidate
graph files, going back a couple years. We had to choose from about 14
different files, so naturally we put these filenames into a markdown
table and added a dedicated column into which we started adding metadata
tags describing the file’s suitability as a proof-of-concept test
subject for this experiment. 

A couple “wish-we-had-it”’s sprung from this:
- The list of filenames came from an unversioned text file. It would
  be nice to be able to generate this same list of names as a “select”
  operation on our new markdown table.
- It would be nice to be able to filter by tags, the exact sort of
  intended use-case of the “filter-by-tags” operation in data-pipes.
- We have a “created-at” date-time column. It would be nice to
  sort by this, perhaps after the filter was applied.

Immediately it became clear that this use-case encompassed the full,
original vision for “data-pipes”: Arbitrarily composable, stream-centric
collection operations (basically an ersatz, text-file-based RDBMS).

The last couple months of development have been centered around
realizing this bold new vision:

- start to realize “pho issues” as a layer on top of kiss/data-pipes
- improve cheap arg parse so it can show -N top issues
- overhaul the “kiss-rdb” API (& SA’s) to honor resource management
- same to dismantle the “collection implementation” monolith
- enhance the data-pipes operations to be composable in two ways:
  as soft pipelines or hard pipelines

This commit achieves (the beginnings of) the last item on that list.
From here, we can attempt the unification of documentation / realize
the isomorphism of graphs-from-markdown tables, and also get our issue
tracking to be more robust and usable, and try to knock out all our
remaining open issues towards some actual substantial application of
our toolkits.

(times)

  10-30 19:07  personal: begin editing video
  10-31 12:23  ok finished exercise / tiktok explosion
        16:20  didn't get that far. took nap instead. resumed.
        20:56  we have parsed the compound command, now to execute
  11-01 08:55  resume with implementation of chain (now called pipeline)
  11-02 09:07  yesterday was a dud. today will be different
  11-02 21:32  it's going well, i need a dance break
               (we got our first compound soft pipeline to work)
  11-05 00:18  begin assembling commit
        10:05  now
.