docs: merge two roadmaps, re-house one as markdown table ((1375))
Discussion:

Let’s review some stuff we discussed in the last commit: We used to
keep a non-versioned file in `z/` called “ROADMAPS”. It was a flat
list of filenames that we would refer to when we were looking for one
of our roadmap graph-viz files.

In this commit we add this new file “roadmaps.md” and with our shiny
new [dp] tools we can duplicate the functionality of having that
plaintext list (with `select`).

Again as discussed at length in a nearby previous commit, an
over-arching goal is to get away from the balkanized habit of using
the mash-mash of plain text files, markdown files and dot files IFF
just a markdown table (file) would suffice.

Broadly our punchlist here was:
1) get insert issue to work just right
2) make sure graph viz generation meets your requirements
3) execute unifying 2 filesystem files and 1 real whiteboard!

In this commit, we were happy with (1), (2) was good enough (and we
fought off the urge to make it perfect) and for (3), we re-absorbed
one tiny graph, we re-housed the content of another graph into the markdown table format. What remains is 👉 another filesystem file,
and a real life physical whiteboard…

(note on times: we took it easy for a few days
during this transitional time)
.