feat(pho): begin to cover issues CLI ((1384))
Exit criteria: be able to open issues with the CLI

Discussion: Our Bold Vision for “issues” from about a month ago was
that it would be a thin layer on top of the “markdown table” storage
adapter. It was with hubris that we assumed we could sidestep testing
the CLI because it would be a thin layer on top of our application
function layer (API), which in turn was supposed to be a thin layer
on top of the kiss storage.

But of course, in practice it didn’t work out that way. Our work in
the last commit (making diff lines a part of the result structure) is
“good” but it introduced breaking changes that didn’t get caught by
our tests (we sort of alluded to this in the last commit message).

Because going forward we want to lean more heavily on using “issues”
to open and close issues (and also make graphs), now is the time to
up the coverage for that one.


Other things

- Unify some things around end states

- Again it was with overzealous kiss-y-ness that we thought that
  “enver” was a suitable use for the fifth parameter of CLI’s.

- Dependencies: some of these (numpy) might just be for stuff that’s
  not in our main development flow and might not really be necessary.
  prune dependencies later!

Some of these 

(times)
  11-12 12:34  begin looking into giving better coverage to top/list
        17:59  break for dance (then dileta set then maybe zz early)
  11-13 12:51  nap for like 10 min
        14:56  now
.