feat(pho issues graph): make “show group nodes” opt-in (+2=30)
Discussion: Initially, to model the reality of a node being inside a
subgraph, we necessitated that you model a specific issue (row and
identifier) modeling the group itself. In the render, we had both the
participating nodes be *inside* the subgraph (rectangle) and also arcs
pointing to this one node that represented the group. This second
behavior wasn’t how our target graphs looked, but it was easier to
implement and we weren’t sure how we wanted to specify this behavior.

Now, we we are pretty sure that we *don’t* want these arcs, we just
want the nodes to be inside the rectangle and that’s it. It appears
that these extra arcs and this extra node are purely extraneous, because
the information contained in them is redundant with the existence of
nodes inside the subgraph (rectangle), and its label.

Down the road there may be issues with this new default behavior, if
you have other notes outside of the subgraph pointing to the node that
represents the group itself. You would probably end up with nodes
pointing to just a graph-viz node identifier… So it might require some
foresight for you to know how you want to model your issues in the table
based on how you want your graph to look; and that is okay.

(times)
  11-13 20:49  discovered curses
        21:04  ok switch gears: absorb one graph into the other
  11-14 12:54  resume absorb thing into thing
        16:34  woke from nap. begin
        19:55  now
.