refactor(microservice-lib): IRC-style not fixed-width headers (6)
Explanation: From our perspective, the most mentally challenging part of
TCP/IP is getting used to the constraints of blocking I/O operations:

When you go to read a tcp/ip message you must specify how many bytes
you are reading; and if that many bytes isn’t available yet, it will
block until that many bytes show up (or forever if they never show up).

(Indeed this same dyanamic can be at play in a lot of I/O operations,
like reading from STDIN.)

In the example (from the python source code comments?), the server keeps
reading from the client until the client closes the connection (and
`recv` results in a zero-length bytestring). The client knows exactly
how many bytes to expect from the server because of the contrived
behavior of the “echo” server: the length of the response will always
equal the length of the request.

In the case of client reading the server response, the read operation
blocks until all the N bytes are available. The server knows when to
stop reading from client because the client closes the connection when
it’s done sending.

Our requirements differ from this overly-simple example because A)
we want a single client connection to be able to send multiple requests
to a server (in sequence), and B) we want the messages (request and
response) to be of any arbitrary length, not some length pre-determined.

In the state of things way before this commit (1 or 2 years ago), we
overcame this limitation with the `select` function. With this function,
you provide a timeout amount and it returns without blocking after that
time is elapsed, even if no new bytes were available.

Upon revisiting this dynamic, we decided that this was a sub-optimal
solution because it doesn’t accommodate the possibility of multipart
messages; that is, it may be the case that more bytes aren’t available
because of network latency, and not because that’s the end of the
message. That’s why, this time around, we started thinking about headers:

In the state of things before this commit, we overcame this limitation
by using “fixed width headers”: a “header line” was exactly 80 bytes
wide, and there was one header line for expressing the “content length”
of the “payload” of the message. Provided that clients and servers both
employ this API correctly, they could read headers from each other
without forever-blocking because they could know it’s safe to request
the next 80 bytes from the message (until they reached the end of the
headers, which itself was signified by a fixed-width header line that
said something like “End of headers” lol).

The really cool part was, this worked. But after discussion with a
colleague, we concluded that we could overcome the inelegant clunkiness
of fixed-width headers by reading the tcp/ip message byte-by-byte with
the assumption that every header line ends in a newline (and that the
end of the headers is indicated by a single blank line) (Probably
we are reinventing a very small part of HTTP that we read somewhere
24 years ago.)

What’s really *really* cool is that *this* worked. The only problem
*now* is that our heavily-mocked tcp/ip connections underwent a lot of
strain to be able to accommodate this byte-by-byte reading in a way that
wasn’t obscenely wasteful or effectively unusable from the definition
side.

(times)
  23:42  begin big refactor
  00:51  finished DRY encode decode and print hack.
         begin change how headers are parsed
  01:04  ok really about to start coding the new headers parsing thing
  01:56  new way of headers is sketched out. but need a rest
  10:21  resume. 21 minutes later visual test passing.
         tests will be annoying.
  11:43  finished catching up tests, maybe
.