test(curses-yikes): introduce auto-generating buttons (32)
Discussion: The way we “injected” state stuff into some of the
components was kind of a mess. Now, all concrete component classes
“know” of this idea of state, even those that are stateless. And
the relationship between a component class and its formal FSA is
more obvious now when you look at the code.

now, the “changes” description language addresses its intended
controller first (there’s already an input controller, a buttons
controller and a selection controller).

(addendum: we initially "made" this commit at the time indicated below
but we didn't actually hit the commit button on our gitx, so it was just
sitting there with this mesage (not *this* message) in its memory for
a few hours while we moved on to the next work. Furthermore, we did
several "amend" commits to what we thought was this commit; but was
actually the previous commit. We just untangled this just now, but
anyway the point is, if you notice a descrepancy between timetracking
times and this commit time, this is why.)

(times)
  11-27 01:06  begin to try and imagine how FSAs will work with buttons
        02:33  finished writing long pseudocode algo, now writing test
        05:19  we have cool index. zz
        09:49  resume
        16:27  change page works and shows buttons generated from FSAs
               (now we have to get the input controller to change the
               button page, as well as manage a redraw hash)
        19:53  resume
  11-28 01:48  resumed after nap after run. implement apply changes
        09:20  begin writing first transition action (toggle)
        10:12  now
.