step(pho): snapshot of pelican project using alabaster theme (ROUGH)
Discussion: A crucial intermediate step between snapshotting
sphinx-generated html and cutting it up to be a pelican theme IS
beautifying the html so we don’t lose our mind cutting parts out.

This is obscenely rough: the html calls for a bunch of javascript
that we don’t have and won’t load for now. The only purpose of this
snapshot is to serve as a safe waypoint where the output looks right
(even though it’s functionally non-functional in several senses).

we used https://www.freeformatter.com/html-formatter.html lol

We added a thing in our pelican config file (“template”) for detecting
when a special environment variable is set pointing to the pelican
theme to use. “Normally” developers would probably write the filesystem
path directly in the config but we take the heroku approach of keeping
such things out of versioned config.

$echo "$PHO_THEME_PATH"  # =>
  [mono-repo-dir]/pho-themes/for-pelican/alabaster-for-pelican

We added an option to the task that makes the sphinx project to create
the project in an ordinary directory which it leaves there, rather than
creating it in a temp dir and moving the needed files out. This new
option is useful for doing a side-by-side comparison of a minimal
sphinx-generated site with the pelican-generated site generated at this
HEAD. Visually we confirmed these two sites look the same at this moment.
.