test(pho): begin “watch” command ((1475))
Discussion: Create our first watch adapter.

This doesn’t get as far as sending the proper message over the TCP/IP
wire to the server, but it’s a big enough sweep of work that we wanted
to commit an intermediate snapshot.

Spike our rough sketch of how we envisioned this going with our
use of dictionaries and JSON to wrap up function-call-like requests
and then responses (RPC lol).

(times)
  18:31  begin make a rough prototype of an fswatch thing
  21:43  woke from nap, make version check for watchexec
  23:22  finished way over-engineered version check of watchexec
  00:46  happy with the watch command.
         about to flesh-out the “connect” command a lot more
  04:13  did great, pinging server
  12:28  now
.